requirements.txt
requests
openai 
python-dotenv 
faiss-cpu 
numpy 
tiktoken
streamlit
pymupdf
python-docx
openpyxl